Fixed C# parser bug regarding integer literals between long.MaxValue + 1 and ulong.MaxValue. C# spec 2.4.4.2: Implemented rules regarding int.MinValue and long.MinValue (they get tokenized to Token.Minus + Token.Literal(uint/ulong), but must be parsed as signed int/long)
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/branches/2.1@2638 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Libraries/NRefactory/Project/Src/Lexer/CSharp/Lexer.cs
@@ -321,7 +321,7 @@ namespace ICSharpCode.NRefactory.Parser.CSharp
 				}
 				
 				// Try to determine a parsable value using ranges. (Quick hack!)
-				double d = 0;
+				decimal d = 0;
 				if (ishex) {
 					ulong result;
 					if (ulong.TryParse(digit, NumberStyles.HexNumber, null, out result)) {
@@ -331,7 +331,7 @@ namespace ICSharpCode.NRefactory.Parser.CSharp
 						return new Token(Tokens.Literal, x, y, stringValue.ToString(), 0);
 					}
 				} else {
-					if (!Double.TryParse(digit, NumberStyles.Integer, null, out d)) {
+					if (!decimal.TryParse(digit, NumberStyles.Integer, null, out d)) {
 						errors.Error(y, x, String.Format("Can't parse integral constant {0}", digit));
 						return new Token(Tokens.Literal, x, y, stringValue.ToString(), 0);
 					}
